XIVY-2942 refactor: move data, detail and master to src
