+++ packages/dataclass-editor/src/data/validation-utils.ts
