+++ packages/dataclass-editor/src/detail/AnnotationsTable.tsx
