+++ packages/dataclass-editor/src/detail/dataclass/DataClassDetailContent.tsx
@@ -1,5 +1,5 @@
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger, Flex } from '@axonivy/ui-components';
-import { useAppContext } from '../../../../context/AppContext';
+import { useAppContext } from '../../context/AppContext';
 import { useDataClassChangeHandlers } from '../../data/dataclass-change-handlers';
 import { isEntity } from '../../data/dataclass-utils';
 import { AnnotationsTable } from '../AnnotationsTable';
