+++ packages/dataclass-editor/src/index.ts
@@ -1,4 +1,4 @@
-export * from './components/dataclass/data/dataclass';
+export * from './data/dataclass';
 export { default as DataClassEditor } from './DataClassEditor';
 export * from './protocol/client-json-rpc';
 export * from './protocol/ClientContextProvider';
