+++ packages/dataclass-editor/src/master/DataClassMasterContent.css
